test(demo): widen the RMS accumulation before the multiply (CodeQL alert 207)
One-line cast matching the tree's RMS idiom (client_eq_test.cpp:53 et
al.). Numerically inert here — 24-sample blocks of ±1.0-bounded audio —
applied to clear the code-scanning gate.

+++ tests/noise_mixer_test.cpp
@@ -190,7 +190,7 @@ void testCwKeysDecodableMorse()
     QVector<double> env;
     for (int i = 0; i + blk <= audio.size(); i += blk) {
         double acc = 0.0;
-        for (int j = 0; j < blk; ++j) acc += audio[i + j] * audio[i + j];
+        for (int j = 0; j < blk; ++j) acc += static_cast<double>(audio[i + j]) * audio[i + j];
         env.append(std::sqrt(acc / blk));
     }
     double emax = 0.0;
